+++ src/Classifiers/KNearestNeighbors.php
@@ -235,6 +235,7 @@ class KNearestNeighbors implements Estimator, Learner, Online, Probabilistic, Pe
             $weights = array_count_values($labels);
         }
 
+        /** @var array<string,float|int> $weights */
         return argmax($weights);
     }